refactor: extract grid utilities
